update code render

+++ celeryworker/tasks.py
@@ -82,7 +82,7 @@ def delete_directory(video_id):
 def task_failure_handler(sender, task_id, exception, args, kwargs, traceback, einfo, **kw):
     video_id = args[0].get('video_id')
     worker_id = "None"
-    update_status_video("Render Lỗi : Xử Lý Video Không Thành Công!", video_id, task_id, worker_id)
+    update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Xử Lý Video Không Thành Công!", video_id, task_id, worker_id)
     delete_directory(video_id)
 # Xử lý khi task bị hủy
 
@@ -106,10 +106,14 @@ def render_video(self, data):
     
     update_status_video("Đang Render : Đang xử lý video render", data['video_id'], task_id, worker_id)
     success = create_or_reset_directory(f'media/{video_id}')
+    
+    if not os.path.exists("video")  and not os.path.exists("video_screen") :
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Thiếu các tệp video  và  video_screen ", data['video_id'], task_id, worker_id)
+        return
 
     if not success:
         shutil.rmtree(f'media/{video_id}')
-        update_status_video("Render Lỗi : Không thể tạo thư mục", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tạo thư mục", data['video_id'], task_id, worker_id)
         return
     update_status_video("Đang Render : Tạo thư mục thành công", data['video_id'], task_id, worker_id)
 
@@ -117,7 +121,7 @@ def render_video(self, data):
     success = download_image(data, task_id, worker_id)
     if not success:
         shutil.rmtree(f'media/{video_id}')
-        update_status_video("Render Lỗi : Không thể tải xuống hình ảnh", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tải xuống hình ảnh", data['video_id'], task_id, worker_id)
         return
     update_status_video("Đang Render : Tải xuống hình ảnh thành công", data['video_id'], task_id, worker_id)
     #THử
@@ -133,7 +137,7 @@ def render_video(self, data):
     success = merge_audio_video(data, task_id, worker_id)
     if not success:
         shutil.rmtree(f'media/{video_id}')
-        update_status_video("Render Lỗi : Không thể nối giọng đọc và chèn nhạc nền", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể nối giọng đọc và chèn nhạc nền", data['video_id'], task_id, worker_id)
         return
     
     update_status_video("Đang Render : Nối giọng đọc và chèn nhạc nền thành công", data['video_id'], task_id, worker_id)
@@ -141,7 +145,7 @@ def render_video(self, data):
     # Tạo video
     success = create_video_lines(data, task_id, worker_id)
     if not success:
-        # shutil.rmtree(f'media/{video_id}')
+        shutil.rmtree(f'media/{video_id}')
         return
    
     # Tạo phụ đề cho video
@@ -149,18 +153,18 @@ def render_video(self, data):
     if not success:
         shutil.rmtree(f'media/{video_id}')
         return
-    
     # Tạo file
     success = create_video_file(data, task_id, worker_id)
     if not success:
-        # shutil.rmtree(f'media/{video_id}')
+        shutil.rmtree(f'media/{video_id}')
         return
-
+    
     success = upload_video(data, task_id, worker_id)
     if not success:
-        update_status_video("Render Lỗi : Không thể upload video", data['video_id'], task_id, worker_id)
+        shutil.rmtree(f'media/{video_id}')
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể upload video", data['video_id'], task_id, worker_id)
         return
-    # shutil.rmtree(f'media/{video_id}')
+    shutil.rmtree(f'media/{video_id}')
     update_status_video(f"Render Thành Công : Đang Chờ Upload lên Kênh", data['video_id'], task_id, worker_id)
 
 @shared_task(bind=True, priority=1,name='render_video_reupload',time_limit=140000,queue='render_video_reupload')
@@ -170,6 +174,10 @@ def render_video_reupload(self, data):
     video_id = data.get('video_id')
     # Kiểm tra xem task có bị hủy không ngay từ đầu
     update_status_video("Đang Render : Đang xử lý video render", data['video_id'], task_id, worker_id)
+    
+    if not os.path.exists("video")  and not os.path.exists("video_screen") :
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Thiếu các tệp video  và  video_screen ", data['video_id'], task_id, worker_id)
+        return
     
     success = create_or_reset_directory(f'media/{video_id}')
     if not success:
@@ -476,11 +484,11 @@ def upload_video(data, task_id, worker_id):
             return True
         
         except Exception as e:
-            update_status_video(f"Render Lỗi : Lỗi khi upload file: {str(e)}", video_id, task_id, worker_id)
+            update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Lỗi khi upload file: {str(e)}", video_id, task_id, worker_id)
             logging.error(f"Error during file upload: {str(e)}")
             return False
     else:
-        update_status_video(f"Render Lỗi : Google API credentials error", video_id, task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Google API credentials error", video_id, task_id, worker_id)
         logging.error("Google API credentials error")
         return False
 
@@ -535,7 +543,7 @@ def create_video_file(data, task_id, worker_id):
                     update_status_video(f"Đang Render: Đã xuất video {percentage:.2f}%", video_id, task_id, worker_id)
                 except Exception as e:
                     print(f"Error parsing time: {e}")
-                    update_status_video("Render Lỗi : Không thể tính toán hoàn thành", data['video_id'], task_id, worker_id)
+                    update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tính toán hoàn thành", data['video_id'], task_id, worker_id)
         process.wait()
         
     if process.returncode != 0:
@@ -543,10 +551,10 @@ def create_video_file(data, task_id, worker_id):
             print("FFmpeg encountered an error.")
             stderr_output = ''.join(process.stderr)
             print(f"Error log:\n{stderr_output}")
-            update_status_video("Render Lỗi : không thể render video hoàn thành ", data['video_id'], task_id, worker_id)
+            update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   không thể render video hoàn thành ", data['video_id'], task_id, worker_id)
         except:
             pass
-        update_status_video("Render Lỗi : không thể render video hoàn thành ", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   không thể render video hoàn thành ", data['video_id'], task_id, worker_id)
         return False
     else:
         print("Lồng nhạc nền thành công.")
@@ -706,7 +714,7 @@ def create_subtitles(data, task_id, worker_id):
             update_status_video("Đang Render : Tạo phụ đề thành công", data['video_id'], task_id, worker_id)
             return True
     except:
-        update_status_video("Render Lỗi : Không thể tạo phụ đề", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tạo phụ đề", data['video_id'], task_id, worker_id)
         return False
 
 def merge_audio_video(data, task_id, worker_id):
@@ -851,7 +859,7 @@ def cut_and_scale_video_random(input_video, output_video, duration, scale_width,
                 "-i",base_video,  # Video overlay
                 "-t", str(duration),                         # Lấy 20 giây cho video overlay
                 "-filter_complex", 
-                f"[0:v]fps=24,scale={scale_width}:{scale_height},setpts={scale_factor}*PTS[bg];[1:v]scale={scale_width}:{scale_height}[fg];[bg][fg]overlay=format=auto[outv]",  # Bộ lọc phức tạp
+                f"[0:v]fps=24,scale={scale_width}:{scale_height},fps=24,setpts={scale_factor}*PTS[bg];[1:v]scale={scale_width}:{scale_height}[fg];[bg][fg]overlay=format=auto[outv]",  # Bộ lọc phức tạp
                 "-map", "[outv]",                   # Ánh xạ đầu ra video
                 "-r", "24",                         # Tốc độ khung hình đầu ra 24 fps
                 "-c:v", "h264_nvenc",               # Codec video H.264 với NVIDIA NVENC
@@ -865,7 +873,7 @@ def cut_and_scale_video_random(input_video, output_video, duration, scale_width,
             "-i", input_video,
             "-ss", start_time_str,   # Thời gian bắt đầu cắt của video
             "-t", str(duration),     # Thời gian video cần cắt
-            "-vf", f"scale_cuda={scale_width}:{scale_height},setpts={scale_factor}*PTS",  # Thay đổi độ phân giải và tốc độ video
+            "-vf", f"fps=24,scale_cuda={scale_width}:{scale_height},setpts={scale_factor}*PTS",  # Thay đổi độ phân giải và tốc độ video
             "-r", "24",              # Tốc độ khung hình đầu ra
             "-c:v", "hevc_nvenc",    # Codec video
             '-preset', 'fast',   
@@ -963,7 +971,7 @@ def process_video_segment(data, text_entry, data_sub, i, video_id, task_id, work
         # Kiểm tra nếu thời lượng âm thanh không hợp lệ
         if duration <= 0:
             update_status_video(
-                        f"Render Lỗi : Thời lượng âm thanh không hợp lệ",
+                        f"Render Lỗi :  {os.getenv('name_woker')}   Thời lượng âm thanh không hợp lệ",
                         video_id, task_id, worker_id
                     )
             
@@ -976,7 +984,7 @@ def process_video_segment(data, text_entry, data_sub, i, video_id, task_id, work
         # Kiểm tra đường dẫn file
         if not file:
             update_status_video(
-                        f"Render Lỗi : Đường dẫn url không hợp lệ",
+                        f"Render Lỗi :  {os.getenv('name_woker')}   Đường dẫn url không hợp lệ",
                         video_id, task_id, worker_id
                     )
             raise FileNotFoundError(f"File not found for URL: {text_entry.get('url_video')}")
@@ -989,7 +997,7 @@ def process_video_segment(data, text_entry, data_sub, i, video_id, task_id, work
         file_type = check_file_type(path_file)
         if file_type not in ["video", "image"]:
             update_status_video(
-                        f"Render Lỗi : Loại file không hợp lệ",
+                        f"Render Lỗi :  {os.getenv('name_woker')}   Loại file không hợp lệ",
                         video_id, task_id, worker_id
                     )
             raise ValueError(f"Unsupported file type: {file_type} for {path_file}")
@@ -1073,7 +1081,7 @@ def image_to_video_zoom_in(image_file,path_video, duration,scale_width, scale_he
             '-i', base_video,                            # Video overlay                          # File âm thanh
             '-filter_complex',
             f"[0:v]format=yuv420p,scale=8000:-1,zoompan=z='zoom+0.005':x=iw/2-(iw/zoom/2):y=ih/2-(ih/zoom/2):d={duration}*24:s={scale_width}x{scale_height}:fps=24[bg];"
-            f"[1:v]scale={scale_width}:{scale_height}[overlay_scaled];"
+            f"[1:v]fps=24,scale={scale_width}:{scale_height}[overlay_scaled];"
             f"[bg][overlay_scaled]overlay=format=auto,format=yuv420p[outv]",
             '-map', '[outv]',                            # Lấy video đã xử lý
             '-t', time_video,                            # Đặt thời lượng video
@@ -1484,7 +1492,7 @@ def download_audio(data, task_id, worker_id):
                         result = future.result()  # Lấy kết quả từ công việc hoàn thành
                         if result[0] is False:  # Nếu có lỗi trong quá trình tải
                             print("Lỗi khi tải giọng nói, dừng toàn bộ tiến trình.")
-                            update_status_video("Render Lỗi : Lỗi khi tải giọng nói, dừng toàn bộ tiến trình.", data['video_id'], task_id, worker_id)
+                            update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Lỗi khi tải giọng nói, dừng toàn bộ tiến trình.", data['video_id'], task_id, worker_id)
                             # Hủy tất cả các công việc chưa hoàn thành
                             for f in futures.keys():
                                 f.cancel()
@@ -1500,13 +1508,13 @@ def download_audio(data, task_id, worker_id):
                     except Exception as e:
                         print(f"Lỗi khi xử lý giọng đọc cho đoạn văn bản {text_entries[idx]['id']}: {e}")
                         update_status_video(
-                            f"Render Lỗi : Lỗi khi tạo giọng đọc - {e}",
+                            f"Render Lỗi :  {os.getenv('name_woker')}   Lỗi khi tạo giọng đọc - {e}",
                             video_id, task_id, worker_id
                         )
                         # Hủy tất cả các công việc chưa hoàn thành
                         for f in futures.keys():
                             f.cancel()
-                        update_status_video("Render Lỗi : Không thể tải xuống âm thanh", data['video_id'], task_id, worker_id)
+                        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tải xuống âm thanh", data['video_id'], task_id, worker_id)
                         return False  # Dừng toàn bộ nếu gặp lỗi
                 # Ghi vào input_files.txt theo đúng thứ tự ban đầu của text_entries
                 for file_name in result_files:
@@ -1514,7 +1522,7 @@ def download_audio(data, task_id, worker_id):
                         file.write(f"file 'voice/{os.path.basename(file_name)}'\n")
         return True
     except Exception as e:
-        update_status_video("Render Lỗi : Không thể tải xuống âm thanh", data['video_id'], task_id, worker_id)
+        update_status_video(f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tải xuống âm thanh", data['video_id'], task_id, worker_id)
         return False
 
 def format_timestamp(seconds):
@@ -1814,7 +1822,7 @@ def download_single_image(url, local_directory):
         return True  # Trả về True nếu tệp đã tồn tại
 
     print(f"Đang tải xuống hình ảnh từ: {url}")
-    for attempt in range(30):  # Thử tải lại 30 lần nếu thất bại
+    for attempt in range(5):  # Thử tải lại 30 lần nếu thất bại
         try:
             response = requests.get(url, stream=True, timeout=200)
             if response.status_code == 200:
@@ -1850,7 +1858,7 @@ def download_image(data, task_id, worker_id):
     for iteam in text_entries:
         if iteam.get('url_video') =="":
             update_status_video(
-                        f"Render Lỗi : iteam hình ảnh lỗi vui lòng xử lý lại",
+                        f"Render Lỗi :  {os.getenv('name_woker')}   iteam hình ảnh lỗi vui lòng xử lý lại",
                         video_id, task_id, worker_id
                     )
             return False
@@ -1886,7 +1894,7 @@ def download_image(data, task_id, worker_id):
                 else:
                     # Hủy tất cả các tác vụ còn lại khi gặp lỗi tải xuống
                     update_status_video(
-                        f"Render Lỗi : Không thể tải xuống hình ảnh -{url}",
+                        f"Render Lỗi :  {os.getenv('name_woker')}   Không thể tải xuống hình ảnh -{url}",
                         video_id, task_id, worker_id
                     )
                     for pending in future_to_url:
@@ -1895,7 +1903,7 @@ def download_image(data, task_id, worker_id):
             except Exception as e:
                 print(f"Lỗi khi tải xuống {url}: {e}")
                 update_status_video(
-                    f"Render Lỗi : Lỗi không xác định - {e} - {url}",
+                    f"Render Lỗi :  {os.getenv('name_woker')}   Lỗi không xác định - {e} - {url}",
                     video_id, task_id, worker_id
                 )
                 # Hủy tất cả các tác vụ còn lại và ngừng tiến trình
